fix: 온보딩 페이지 프리렌더링 오류 해결
