+++ src/components/onboarding/onboardingItems/nickName.tsx
